todoapp-backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.0
httpx==0.25.1
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1